requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.2
boto3==1.15.6
botocore==1.18.6
certifi==2020.6.20
chardet==3.0.4
Django==3.1.1
django-cors-headers==3.5.0
django-phonenumber-field==5.0.0
django-storages==1.10.1
djangorestframework==3.12.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
lyricsgenius==2.0.2
pafy==0.5.5
phonenumbers==8.12.10
Pillow==7.2.0
psycopg2==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
pytube==10.5.1
pytz==2020.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tinytag==1.4.0
typing-extensions==3.7.4.3
urllib3==1.25.10
whitenoise==5.2.0
